refactor(user-service): 重构收货地址相关功能
- 修改地址 ID 类型从 int 改为 Long
- 增加参数验证和用户身份验证
-优化删除地址逻辑，处理不存在或无权限访问的情况
- 新增地址参数验证逻辑
- 更新相关枚举和异常处理

+++ shop-user-service/src/main/java/fun/timu/shop/user/controller/AddressController.java
@@ -48,7 +48,7 @@ public class AddressController {
      * @return
      */
     @DeleteMapping("/del/{address_id}")
-    public JsonData del(@PathVariable("address_id") int addressId) {
+    public JsonData del(@PathVariable("address_id") Long addressId) {
         return addressService.del(addressId);
     }
